src/requirements.txt
dj-database-url==0.4.1
dj-static==0.0.6
django-toolbelt==0.0.1
django-treebeard==4.0.1
djangorestframework==3.4.6
gunicorn==19.6.0
psycopg2==2.6.2
